x86/boot: Convert move_memory() to use bootstrap_map_addr()
move_memory() is very complicated, and buggy.  In order to fix the latter, we
have to address the former.

Given prior cleanup, bootstrap_map() is now implemented in terms of
bootstrap_map_addr(), meaining that it is counterproductive to plumb the
mapping through module_t.

Delete mod, and introduce two same-sized/named fields.  At this point in boot,
neither fields have their named purpose, so indicate the purpose in comments.

No functional change.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
---
CC: Jan Beulich <JBeulich@suse.com>
CC: Roger Pau Monné <roger.pau@citrix.com>
CC: Daniel P. Smith <dpsmith@apertussolutions.com>

This is broken out both to ease review (I recommend git diff --color-words),
and for bisection purposes given how many times I've failed to edit this logic
correctly.

This does not conflict with anything in Hyperlaunch v7, but it does remove the
only callers of bootstrap_map() remaining by the end of the series, allowing
for even more cleanup.

In terms of bugs.  The ones I've spotted so far are:

 * The blksz calculation depends on the previous thing having been unmapped
   first.

 * The calculation halving blksz based on BOOTSTRAP_MAP_BASE is bogus.  It
   might plausibly be a remnent of 32bit Xen with the absence of highmem.

 * The caller of move_memory() is strictly moving modules forward in
   memory (dst > src) with a possibility of partial overlap.  The while loop
   maps dst and src, blksz chunks at a time, ascending, with s < d.  This
   results in memmove() doing an unconditional backwards copy.

   I don't see how this logic could ever have worked for an overlapping move.

 * The caller of move_memory() already has a good mapping for dst, so we don't
   even need to split BOOTSTRAP_MAP_LIMIT in half to map both parts.