x86/boot: Simplify size calculations in move_memory()
While both src and dst are similar, src is mapped only accounting for src's
size, while dst is mapped based on the minimum of both.  This means that in
some cases, an overly large mapping is requested for src.

Rework the sz calcuation to be symmetric, and leave an explanation of how
logic works.

No functional change.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
---
CC: Jan Beulich <JBeulich@suse.com>
CC: Roger Pau Monné <roger.pau@citrix.com>
CC: Daniel P. Smith <dpsmith@apertussolutions.com>

And as if by magic, all transient fields from converting module_t are gone,
with bloat-o-meter reporting:

  add/remove: 0/0 grow/shrink: 0/1 up/down: 0/-74 (-74)
  Function                                     old     new   delta
  move_memory                                  265     191     -74

which is all removal of logic inside the while loop.