x86: Initialize mapcache for PV, HVM, and idle domains
To support the transition away from the direct map, the mapcache will now
be used by HVM and idle domains as well. This patch lifts the `mapcache`
to the arch level and moves its initialization earlier in
`arch_domain_create()` to cover PV, HVM, and idle domains.

For the idle domain to utilize the mapcache, this patch also populates the
mapcache page tables within the `PERDOMAIN` region and adjusts the
initialization sequence.

With this change, mapcache initialization is now unified across all domain
types—PV, HVM, and idle.

Signed-off-by: Wei Liu <wei.liu2@citrix.com>
Signed-off-by: Wei Wang <wawei@amazon.de>
Signed-off-by: Hongyan Xia <hongyxia@amazon.com>
Signed-off-by: Julien Grall <jgrall@amazon.com>
Signed-off-by: Elias El Yandouzi <eliasely@amazon.com>

----

	Changes in V4:
        * Reword the commit message
        * Rebase it on top of staging
            * The logic for the creation of the domain has been reworked
              so introduced #ifdef CONFIG_X86 in the common code to
              initialise the mapcache

Changes in V2:
        * Free resources if mapcache initialisation fails
        * Remove `is_idle_domain()` check from `create_perdomain_mappings()`