xen/lib: Introduce SHA256
Future changes will need to calculate SHA256 digests.  Introduce an
implementation in lib/, derived from Trenchboot which itself is derived from
Linux.

The implementation is optimised for size, and for Xen's expected usecase of
only needing to digest single messages.  In order to be useful to other
architectures, it is careful with endianness and misaligned accesses as well
as being more MISRA friendly, but is only wired up for x86 in the short term.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
---
I've compiled it for all archtiectures, but not done a gitlab run yet with
selftests enabled.