x86/traps: Avoid OoB accesses to print the data selectors
_show_registers() prints the data selectors from struct cpu_user_regs, but
these fields are sometimes out-of-bounds.  See commit 6065a05adf15
("x86/traps: 'Fix' safety of read_registers() in #DF path").

There are 3 callers of _show_registers():

 1. vcpu_show_registers(), which always operates on a scheduled-out vCPU,
    where v->arch.user_regs (or aux_regs on the stack) is always in-bounds.

 2. show_registers() where regs is always an on-stack frame.  regs is copied
    into a local variable first (which is an OoB read for constructs such as
    WARN()), before being modified (so no OoB write).

 3. do_double_fault(), where regs is adjacent to the stack guard page, and
    written into directly.  This is an out of bounds read and write, with a
    bodge to avoid the writes hitting the guard page.

Furthermore, these fields are a vestigial remenant of vm86 mode, and need to
be changed in order to support FRED.

Therefore, include the data segment selectors in struct extra_state, and use
those fields instead of the fields in regs.  This resolves the OoB write on
the #DF path.  The OoB read in show_registers() is resolved by doing a partial
memcpy() rather than full structure copy.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>