xen: Don't cast away const-ness in vcpu_show_registers()
The final hunk is `(struct vcpu *)v` expressed using a runtime pointer chase
through memory and a technicality of the C type system.

For anyone interested, this is one reason why C cannot optimise any reads
across sequence points, even for a function purporting to take a const object.

Anyway, have the function correctly state that it needs a mutable vcpu.  All
callers have a mutable vCPU to hand, and it removes the runtime pointer chase
in x86.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>